requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sqlalchemy
pydantic
alembic
databases
